+++ lib/chewie.dart
@@ -2,6 +2,6 @@ library chewie;
 
 export 'src/chewie_player.dart';
 export 'src/chewie_progress_colors.dart';
-export 'src/cupertino_controls.dart';
-export 'src/material_controls.dart';
-export 'src/subtitle_model.dart';
+export 'src/cupertino/cupertino_controls.dart';
+export 'src/material/material_controls.dart';
+export 'src/models/index.dart';