+++ lib/src/models/index.dart
@@ -1,0 +1,2 @@
+export '../material/models/index.dart';
+export 'subtitle_model.dart';