+++ lib/src/notifiers/index.dart
@@ -1,0 +1,1 @@
+export 'player_notifier.dart';